Main-Menu task week 2
Sooo... I finished my tasks. it worked when I tested it. Hopefully it works for everyone else too. When pressing a button a prompt pops-up and asks if you wna to conitnue. If no, it takes you back to the main menu. If you click yes to the play button it loads the game scene. If you click yes on exit, the game closes. The other two buttons have the prompt but no functionality yet as instructed. Is this how you push  change to GitHub?